feat(lemma meaning activity): widen distractor range, reduce lemmas where meaning activity required
