+++ lib/pangea/toolbar/repo/lemma_meaning_activity_generator.dart
@@ -1,16 +1,17 @@
-import 'package:flutter/material.dart';
-
-import 'package:flutter_gen/gen_l10n/l10n.dart';
-
 import 'package:fluffychat/pangea/analytics/enums/construct_type_enum.dart';
+import 'package:fluffychat/pangea/analytics/models/constructs_model.dart';
 import 'package:fluffychat/pangea/analytics/repo/lemma_info_repo.dart';
 import 'package:fluffychat/pangea/analytics/repo/lemma_info_request.dart';
+import 'package:fluffychat/pangea/analytics/repo/lemma_info_response.dart';
 import 'package:fluffychat/pangea/toolbar/enums/activity_type_enum.dart';
 import 'package:fluffychat/pangea/toolbar/models/message_activity_request.dart';
 import 'package:fluffychat/pangea/toolbar/models/multiple_choice_activity_model.dart';
 import 'package:fluffychat/pangea/toolbar/models/practice_activity_model.dart';
+import 'package:fluffychat/widgets/matrix.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/l10n.dart';
 
-class WordMeaningActivityGenerator {
+class LemmaMeaningActivityGenerator {
   Future<MessageActivityResponse> get(
     MessageActivityRequest req,
     BuildContext context,
@@ -33,8 +34,7 @@ class WordMeaningActivityGenerator {
 
     final res = await LemmaInfoRepo.get(lemmaDefReq);
 
-    final choices =
-        LemmaInfoRepo.getDistractorDefinitions(lemmaDefReq.lemma, 3);
+    final choices = await getDistractorMeanings(lemmaDefReq, 3);
 
     if (!choices.contains(res.meaning)) {
       choices.add(res.meaning);
@@ -56,5 +56,48 @@ class WordMeaningActivityGenerator {
         ),
       ),
     );
+  }
+
+  /// From the cache, get a random set of cached definitions that are not for a specific lemma
+  static Future<List<String>> getDistractorMeanings(
+    LemmaInfoRequest req,
+    int count,
+  ) async {
+    final allUses =
+        MatrixState.pangeaController.getAnalytics.constructListModel.uses;
+    final filteredUses = allUses
+        .where((c) => c.lemma != req.lemma && c.category != req.partOfSpeech)
+        .toList();
+    filteredUses.shuffle();
+
+    final List<OneConstructUse> distractorConstructUses =
+        filteredUses.take(count).toList();
+
+    final List<Future<LemmaInfoResponse>> futureDefs = [];
+    for (final construct in distractorConstructUses) {
+      futureDefs.add(
+        LemmaInfoRepo.get(
+          LemmaInfoRequest(
+            lemma: construct.lemma,
+            partOfSpeech: construct.category,
+            lemmaLang: req.lemmaLang,
+            userL1: req.userL1,
+          ),
+        ),
+      );
+    }
+
+    final List<String> distractorDefs = [];
+    for (final def in await Future.wait(futureDefs)) {
+      distractorDefs.add(def.meaning);
+    }
+
+    return distractorDefs;
+  }
+
+  static bool canGenerateDistractors(String lemma, String pos) {
+    return MatrixState.pangeaController.getAnalytics.constructListModel.uses
+        .where((c) => c.lemma != lemma && c.category != pos)
+        .isNotEmpty;
   }
 }
